Renamed Concurrent Dictionary for non PCL support.
