+++ Ext/MsgProcessor.cs
@@ -2,8 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using System.Collections.Concurrent;
-
 using RealtimeFramework.Messaging.Exceptions;
 
 namespace RealtimeFramework.Messaging.Ext {
