+++ Ext/RestWebservice.cs
@@ -1,14 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Net;
-using System.IO;
-using System.Threading.Tasks;
-//using System.Web;
-
+using System;
 using System.Net.Http;
-
 using RealtimeFramework.Messaging.Exceptions;
 
 namespace RealtimeFramework.Messaging.Ext
