+++ Ext/StringExtensions.cs
@@ -1,8 +1,4 @@
-﻿using System.Text.RegularExpressions;
-using System;
-using System.IO;
-using System.Runtime.Serialization.Json;
-using System.Text;
+using System.Text.RegularExpressions;
 
 namespace RealtimeFramework.Messaging.Ext
 {
